Refaktor. Flytt logikk for journalføring av vedtaksbrev fra steg til task. Gjør dette for å kunne ha et felles behandlingsteg for journalføring og distribuering av brev.
