+++ src/main/kotlin/no/nav/tilleggsstonader/sak/brev/JournalførVedtaksbrevSteg.kt
@@ -23,86 +23,13 @@ import org.springframework.stereotype.Service
 import org.springframework.web.client.HttpClientErrorException
 
 @Service
-class JournalførVedtaksbrevSteg(
-    private val brevService: BrevService,
-    private val arbeidsfordelingService: ArbeidsfordelingService,
-    private val journalpostService: JournalpostService,
-    private val brevmottakerRepository: BrevmottakerRepository,
-) : BehandlingSteg<Void?> {
+class JournalførVedtaksbrevSteg : BehandlingSteg<Void?> {
 
-    private val logger = LoggerFactory.getLogger(javaClass)
 
     override fun utførSteg(saksbehandling: Saksbehandling, data: Void?) {
-        val vedtaksbrev = brevService.hentBesluttetBrev(saksbehandling.id)
-        val brevmottaker = hentBrevmottaker(saksbehandling)
 
-        val dokument = Dokument(
-            dokument = vedtaksbrev.beslutterPdf?.bytes ?: error("Mangler beslutterpdf"),
-            filtype = Filtype.PDFA,
-            dokumenttype = utledDokumenttype(saksbehandling),
-            tittel = utledBrevtittel(saksbehandling),
-        )
-
-        val eksternReferanseId = "${saksbehandling.eksternId}-vedtaksbrev"
-
-        val arkviverDokumentRequest = ArkiverDokumentRequest(
-            fnr = saksbehandling.ident,
-            forsøkFerdigstill = true,
-            hoveddokumentvarianter = listOf(dokument),
-            fagsakId = saksbehandling.eksternFagsakId.toString(),
-            journalførendeEnhet = arbeidsfordelingService.hentNavEnhet(saksbehandling.ident)?.enhetId
-                ?: error("Fant ikke arbeidsfordelingsenhet"),
-            eksternReferanseId = eksternReferanseId,
-            avsenderMottaker = lagAvsenderMottaker(brevmottaker),
-        )
-
-        try {
-            val response = journalpostService.opprettJournalpost(arkviverDokumentRequest)
-
-            feilHvisIkke(response.ferdigstilt) {
-                "Journalposten ble ikke ferdigstilt og kan derfor ikke distribueres"
-            }
-
-            brevmottakerRepository.update(brevmottaker.copy(journalpostId = response.journalpostId))
-        } catch (e: HttpClientErrorException) {
-            if (e.statusCode == HttpStatus.CONFLICT) {
-                logger.warn("Konflikt ved arkivering av dokument. Vedtaksbrevet har sannsynligvis allerede blitt arkivert for behandlingId=${saksbehandling.id} med eksternReferanseId=$eksternReferanseId")
-            } else {
-                throw e
-            }
-        }
     }
 
-    private fun lagAvsenderMottaker(brevmottaker: Brevmottaker) = AvsenderMottaker(
-        id = brevmottaker.ident,
-        idType = when (brevmottaker.mottakerType) {
-            MottakerType.PERSON -> BrukerIdType.FNR
-            MottakerType.ORGANISASJON -> BrukerIdType.ORGNR
-        },
-        navn = when (brevmottaker.mottakerType) {
-            MottakerType.PERSON -> null
-            MottakerType.ORGANISASJON -> brevmottaker.navnHosOrganisasjon
-        },
-    )
 
-    private fun hentBrevmottaker(saksbehandling: Saksbehandling): Brevmottaker {
-        return brevmottakerRepository.findByBehandlingId(saksbehandling.id).let {
-            feilHvis(it.size > 1) {
-                "Støtte for flere brevmottakere er ikke implementert"
-            }
-
-            it.first()
-        }
-    }
-
-    private fun utledBrevtittel(saksbehandling: Saksbehandling) = when (saksbehandling.stønadstype) {
-        Stønadstype.BARNETILSYN -> "Vedtak om stønad til tilsyn barn" // TODO
-    }
-
-    private fun utledDokumenttype(saksbehandling: Saksbehandling) =
-        when (saksbehandling.stønadstype) {
-            Stønadstype.BARNETILSYN -> Dokumenttype.BARNETILSYN_VEDTAKSBREV
-        }
-
-    override fun stegType(): StegType = StegType.JOURNALFØRE_VEDTAKSBREV
+    override fun stegType(): StegType = StegType.JOURNALFØR_OG_DISTRIBUER_VEDTAKSBREV
 }
